requirements.txt
altair==4.2.0
apeye==1.2.0
asgiref==3.4.1
astroid==2.6.5
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.1.0
blinker==1.5
boto3==1.20.15
botocore==1.23.15
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.0.0
colorama==0.4.4
comtypes==1.1.11
cron-validator==1.0.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
distlib==0.3.4
Django==3.2.5
django-cors-headers==3.10.0
djangorestframework==3.12.4
domdf-python-tools==3.3.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.2
filelock==3.6.0
Flask==2.1.3
fonttools==4.29.0
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.9.1
ipython==8.0.1
isort==5.9.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==4.14.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
MouseInfo==0.1.3
mypy-extensions==0.4.3
natsort==8.1.0
nest-asyncio==1.5.4
numpy==1.21.4
opencv-python==4.6.0.66
openpyxl==3.0.9
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.5.1
plotly==5.9.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psycopg2-binary==2.9.1
pure-eval==0.2.2
pyarrow==9.0.0
PyAudio==0.2.14
PyAutoGUI==0.9.53
pydeck==0.8.0b1
PyGetWindow==0.0.9
Pygments==2.11.2
pylint==2.9.5
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
pypi-json==0.3.0
pypiwin32==223
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==0.20.0
pyttsx3==2.90
pytweening==1.0.4
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.4
pywin32==303
pyzmq==22.3.0
requests==2.28.1
s3transfer==0.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.1
stack-data==0.2.0
streamlit==1.9.0
tenacity==8.0.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.7
validators==0.20.0
virtualenv==20.14.1
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.1.2
wikipedia==1.4.0
wrapt==1.12.1
zipp==3.8.1
